+++ src/types/index.ts
@@ -58,5 +58,6 @@ export * from './common';
 export * from './exchange';
 export * from './firebase';
 export * from './orders';
+export * from './priceLimit';
 export * from './telegram';
 export * from './telegramCommandHandler';